removing unnecessary junk from offCanvas.
