+++ main.js
@@ -44,45 +44,45 @@ try {
 		bgTexturePath: 'smoke-o.webp',
 	})
 
-	myScene.subscribe(
-		({ modelLoadingProgress }) => {
-			if (modelLoadingProgress === 100) {
-				myScene
-				.setCameraPose({ from: 0, to: 1 })
-				.then(myScene.unLockScroll.bind(myScene))
+	// myScene.subscribe(
+	// 	({ modelLoadingProgress }) => {
+	// 		if (modelLoadingProgress === 100) {
+	// 			myScene
+	// 			.setCameraPose({ from: 0, to: 1 })
+	// 			.then(myScene.unLockScroll.bind(myScene))
 
-				scrollToTop.addEventListener('click', () => {
-					myScene
-					.setScenePose({ from: 4, to: 0, duration: 500, camera: { from: 5, to: 1, rate: 5 }, ignoreCameraCurrentState: true })
-					.then(myScene.unLockScroll.bind(myScene))
-				})
+	// 			scrollToTop.addEventListener('click', () => {
+	// 				myScene
+	// 				.setScenePose({ from: 4, to: 0, duration: 500, camera: { from: 5, to: 1, rate: 5 }, ignoreCameraCurrentState: true })
+	// 				.then(myScene.unLockScroll.bind(myScene))
+	// 			})
 
-			}
-		},
-		['modelLoadingProgress'],
-	)
+	// 		}
+	// 	},
+	// 	['modelLoadingProgress'],
+	// )
 
 
-    myScene.subscribe(({
-		sections, sectionCurrent, cameraTransitionComplete, scrollingStarted,
-	}) => {
+    // myScene.subscribe(({
+	// 	sections, sectionCurrent, cameraTransitionComplete, scrollingStarted,
+	// }) => {
 
-		if (sections) {
+	// 	if (sections) {
 
-		  if (scrollingStarted) {
-			sections.forEach((section) => {
-			  section.style.setProperty('--opacity', '0');
-			  section.classList.remove('active');
-			});
-		  }
+	// 	  if (scrollingStarted) {
+	// 		sections.forEach((section) => {
+	// 		  section.style.setProperty('--opacity', '0');
+	// 		  section.classList.remove('active');
+	// 		});
+	// 	  }
 
-		  if (cameraTransitionComplete) {
-			sections[sectionCurrent].style.setProperty('--opacity', '1');
-			sections[sectionCurrent].classList.add('active');
-		  }
+	// 	  if (cameraTransitionComplete) {
+	// 		sections[sectionCurrent].style.setProperty('--opacity', '1');
+	// 		sections[sectionCurrent].classList.add('active');
+	// 	  }
 
-		}
-	  }, ['sections', 'sectionCurrent', 'cameraTransitionComplete', 'scrollingStarted']);
+	// 	}
+	//   }, ['sections', 'sectionCurrent', 'cameraTransitionComplete', 'scrollingStarted']);
 
 } catch (e) {
 	console.log(e)
